simplify and improve GMS crash UI
GMS now calls GmsCompat directly right before the crash, no need to listen for broadcasts from the
system_server.

Also, improve code that prevents notification spam when GMS chain-crashes.
